requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.35
mysql-connector-python>=9.0.0
pydantic>=2.10.0
pydantic-settings>=2.11.0
python-dotenv>=1.0.1

